Cleanup model download code
